Added netty-tcnative dependency & fixed float-casting problem

+++ src/main/java/com/google/codeu/data/Datastore.java
@@ -62,7 +62,7 @@ public class Datastore {
         UUID id = UUID.fromString(idString);
         String user = (String) entity.getProperty("user");
         String text = (String) entity.getProperty("text");
-	float sentimentScore = (float) entity.getProperty("sentimentScore");
+	      float sentimentScore = ((Double) entity.getProperty("sentimentScore")).floatValue();
         long timestamp = (long) entity.getProperty("timestamp");
 
         Message message = new Message(id, user, text, sentimentScore, timestamp);
@@ -76,7 +76,7 @@ public class Datastore {
 
     return messages;
   }
-  
+
     /** Stores the User in Datastore. */
    public void storeUser(User user) {
     Entity userEntity = new Entity("User", user.getEmail());
@@ -84,13 +84,13 @@ public class Datastore {
     userEntity.setProperty("aboutMe", user.getAboutMe());
     datastore.put(userEntity);
    }
-   
+
    /**
     * Returns the User owned by the email address, or
     * null if no matching User was found.
     */
    public User getUser(String email) {
-   
+
     Query query = new Query("User")
       .setFilter(new Query.FilterPredicate("email", FilterOperator.EQUAL, email));
     PreparedQuery results = datastore.prepare(query);
@@ -98,10 +98,10 @@ public class Datastore {
     if(userEntity == null) {
      return null;
     }
-    
+
     String aboutMe = (String) userEntity.getProperty("aboutMe");
     User user = new User(email, aboutMe);
-    
+
     return user;
    }
